MisPerrisServer/requirements.txt
Django==2.1.7
Pillow==5.4.1
djangorestframework==3.9.1
